+++ pkgs/ffigen/lib/src/code_generator/struct.dart
@@ -45,6 +45,7 @@ class Struct extends Compound {
     super.nativeType,
     super.apiAvailability,
     super.sizeInBytes,
+    super.dependencies,
   });
 
   @override