+++ pkgs/ffigen/lib/src/code_generator/union.dart
@@ -43,6 +43,7 @@ class Union extends Compound {
     super.nativeType,
     super.apiAvailability,
     super.sizeInBytes,
+    super.dependencies,
   });
 
   @override