+++ pkgs/ffigen/test/large_integration_tests/large_test.dart
@@ -202,7 +202,7 @@ void main() {
           ],
           include: (Uri header) => header.pathSegments.last == 'sqlite3.h',
         ),
-        structs: const Structs(dependencies: CompoundDependencies.full),
+        structs: const Structs(),
         visitors: [
           Visitor(
             func: (node) {
@@ -222,6 +222,7 @@ void main() {
               } else {
                 node.isIncluded = true;
               }
+              node.dependencies = CompoundDependencies.full;
             },
             typealias: (node) {
               if (vaRegex.hasMatch(node.originalName)) {